Put request handler implementation details in their own header. refs #70
